Rebased to kernel v6.2, bug fixes
New user pointer sanitizing in the kernel was causing our driver to
crash right away in mk2_read and mk2_write.

Fixed parsing of incoming metadata from the device. We were missing
proper parsing of incoming novation metadata and was prefetching usb
reads based on user count argument, where we really should allocate and
request bigger buffer, because we have to be able to hold metadata as
well.